load_testing/requirements.txt
locust==2.20.0
requests==2.31.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
